pyproject.toml
[project]
name = "active-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyterlab>=4.4.1",
    "mlflow>=3.1.1",
    "modal-python>=0.4.2.1",
    "numpy>=2.2.5",
    "seaborn>=0.13.2",
    "skorch>=0.9.0",
    "torch>=2.7.0",
    "torchvision>=0.22.1",
]
